// prover/src/workers/types.ts
interface BlockProverJob {
    height: number; // epochHeight
    blockIndex: number;
}

interface AggregatorJob {
    height: number;
    index: number;
    left: string;
    right: string;
}

interface SettlementProverJob {
    height: number;
    settlementProofId: string;
}

interface SettlerJob {
    height: number;
}

export type { BlockProverJob, AggregatorJob, SettlementProverJob, SettlerJob };